requirements.txt
pyTelegramBotAPI==3.8.1
requests==2.25.1
Flask==2.0.1
python-dotenv==0.18.0